fix: move default facets to compatiblityLayer
