+++ react/utils/compatibilityLayer.js
@@ -2,6 +2,23 @@
 import { groupBy, pathOr, zipObj } from 'ramda'
 
 import { PATH_SEPARATOR, MAP_VALUES_SEP } from '../constants'
+import { SHIPPING_OPTIONS, SHIPPING_KEY } from './getFilters'
+
+const shippingFacetDefault = {
+  name: SHIPPING_KEY,
+  type: 'DELIVERY',
+  hidden: false,
+  quantity: 0,
+  facets: Object.keys(SHIPPING_OPTIONS).map(option => ({
+    id: null,
+    quantity: 0,
+    name: option,
+    key: SHIPPING_KEY,
+    selected: false,
+    map: SHIPPING_KEY,
+    value: option,
+  })),
+}
 
 export const getMainSearches = (query, map) => {
   const querySegments = (query && query.split(PATH_SEPARATOR)) || []
@@ -79,7 +96,7 @@ export const detachFiltersByType = facets => {
     groupedFilters.TEXT || []
   )
 
-  const deliveries = groupedFilters.DELIVERY || []
+  const deliveries = getFormattedDeliveries(groupedFilters.DELIVERY || [], true)
 
   const categoriesTrees = pathOr(
     [],
@@ -117,5 +134,27 @@ export const buildQueryArgsFromSelectedFacets = selectedFacets => {
       return queryArgs
     },
     { query: '', map: '' }
+  )
+}
+
+const getFormattedDeliveries = (deliveries, showShippingFacet) => {
+  if (!showShippingFacet) {
+    return deliveries.filter(d => d.name !== SHIPPING_KEY)
+  }
+
+  const shippingFacet = deliveries.find(d => d.name === SHIPPING_KEY)
+
+  if (!shippingFacet) {
+    return [...deliveries, shippingFacetDefault]
+  }
+
+  const facetsNotIncluded = shippingFacetDefault.facets.filter(facet =>
+    shippingFacet.facets.every(f => f.value !== facet.value)
+  )
+
+  shippingFacet.facets = [...shippingFacet.facets, ...facetsNotIncluded]
+
+  return deliveries.map(facet =>
+    facet.name === SHIPPING_KEY ? shippingFacet : facet
   )
 }
