+++ react/utils/getFilters.js
@@ -13,26 +13,11 @@ export const SHIPPING_OPTIONS = {
   'pickup-all': 'store/search.filter.shipping.name.pickup-all',
 }
 
+export const SHIPPING_KEY = 'shipping'
+
 const BRANDS_TYPE = 'Brands'
 const PRICE_RANGES_TYPE = 'PriceRanges'
 const SPECIFICATION_FILTERS_TYPE = 'SpecificationFilters'
-const SHIPPING_KEY = 'shipping'
-
-const shippingFacetDefault = {
-  name: SHIPPING_KEY,
-  type: 'DELIVERY',
-  hidden: false,
-  quantity: 0,
-  facets: Object.keys(SHIPPING_OPTIONS).map(option => ({
-    id: null,
-    quantity: 0,
-    name: option,
-    key: SHIPPING_KEY,
-    selected: false,
-    map: SHIPPING_KEY,
-    value: option,
-  })),
-}
 
 const getFilters = ({
   specificationFilters = [],
@@ -41,9 +26,28 @@ const getFilters = ({
   deliveries = [],
   brandsQuantity = 0,
   hiddenFacets = {},
-  showShippingFacet = false,
+  // showShippingFacet = false,
 }) => {
   const intl = useIntl()
+
+  let deliveriesFormatted = deliveries
+
+  const shipping = deliveries.find(d => d.name === SHIPPING_KEY)
+
+  if (shipping) {
+    const shippingFacet = {
+      ...shipping,
+      title: SHIPPING_TITLE,
+      facets: shipping.facets.map(facet => ({
+        ...facet,
+        name: intl.formatMessage({ id: SHIPPING_OPTIONS[facet.name] }),
+      })),
+    }
+
+    deliveriesFormatted = deliveries.map(facet =>
+      facet.name === SHIPPING_KEY ? shippingFacet : facet
+    )
+  }
 
   const hiddenFacetsNames = (
     path(['specificationFilters', 'hiddenFilters'], hiddenFacets) || []
@@ -64,26 +68,6 @@ const getFilters = ({
         }))
     : []
 
-  const deliveriesFormatted = getFormattedDeliveries(
-    deliveries,
-    showShippingFacet
-  )
-
-  const shippingIndex = deliveriesFormatted.findIndex(
-    d => d.name === SHIPPING_KEY
-  )
-
-  if (shippingIndex !== -1) {
-    deliveriesFormatted[shippingIndex] = {
-      ...deliveriesFormatted[shippingIndex],
-      title: SHIPPING_TITLE,
-      facets: deliveriesFormatted[shippingIndex].facets.map(facet => ({
-        ...facet,
-        name: intl.formatMessage({ id: SHIPPING_OPTIONS[facet.name] }),
-      })),
-    }
-  }
-
   return [
     ...deliveriesFormatted,
     ...mappedSpecificationFilters,
@@ -101,28 +85,6 @@ const getFilters = ({
         facets: priceRanges,
       },
   ].filter(Boolean)
-}
-
-const getFormattedDeliveries = (deliveries, showShippingFacet) => {
-  if (!showShippingFacet) {
-    return deliveries.filter(d => d.name !== SHIPPING_KEY)
-  }
-
-  const shippingFacet = deliveries.find(d => d.name === SHIPPING_KEY)
-
-  if (!shippingFacet) {
-    return [...deliveries, shippingFacetDefault]
-  }
-
-  const facetsNotIncluded = shippingFacetDefault.facets.filter(facet =>
-    shippingFacet.facets.every(f => f.value !== facet.value)
-  )
-
-  shippingFacet.facets = [...shippingFacet.facets, ...facetsNotIncluded]
-
-  return deliveries.map(facet =>
-    facet.name === SHIPPING_KEY ? shippingFacet : facet
-  )
 }
 
 export default getFilters
